Fixed up title page
You should be consistent with how your write the names in your committee:
use all initials for all members.

The affiliation for Kerstin Blank was wrong: she is not in Munich (she's in
Potsdam, near Berlin), and "Max Planck Munich" is neither correct nor Dutch.
The full affiliation, including place,
  Max-Planck-Institut für Kolloid- und Grenzflächenforschung, Potsdam
is quite long, and would probably have to be split onto two lines. I think it's
reasonable to drop the place name, but there are other options.

In the "ter verkrijging van..." most (all?) dissertations I've seen the line
breaks are placed differently, and they use archaic "klokke" rather than "om"
for the time -- I don't know if this matters.

You're obviously missing a description of the cover.